+++ utils.js
@@ -89,6 +89,38 @@ function removeBlockComments( text, fileName )
     return text;
 }
 
+function removeLineComments( text, fileName )
+{
+    var result = text.trim();
+
+    if( path.extname( fileName ) === ".jsonc" )
+    {
+        fileName = path.join( path.dirname( fileName ), path.basename( fileName, path.extname( fileName ) ) ) + ".js";
+    }
+
+    var commentPattern;
+    try
+    {
+        commentPattern = commentPatterns( fileName );
+    }
+    catch( e )
+    {
+    }
+
+    if( commentPattern && commentPattern.singleLineComment )
+    {
+        commentPattern.singleLineComment.map( function( comment )
+        {
+            if( result.indexOf( comment.start ) === 0 )
+            {
+                result = result.substr( comment.start.length );
+            }
+        } );
+    }
+
+    return result;
+}
+
 function getTagRegex()
 {
     var c = config.regex();
@@ -254,6 +286,7 @@ module.exports.init = init;
 module.exports.isHexColour = isHexColour;
 module.exports.hexToRgba = hexToRgba;
 module.exports.removeBlockComments = removeBlockComments;
+module.exports.removeLineComments = removeLineComments;
 module.exports.extractTag = extractTag;
 module.exports.getRegexSource = getRegexSource;
 module.exports.getRegexForRipGrep = getRegexForRipGrep;